Check for update code pushed out to CheckForAppUpdate.cs
Check for update code pushed out to CheckForAppUpdate.cs
